\name{GOdata}

\alias{GOdata}

\docType{data}

\title{Example of a topGO data object}
\description{
  This data set contains an instance of a \code{topGOdata} object.
  It can be used to run an enrichment analysis directly.
}

\usage{data(GOdata)}

\source{
  Generated using the \code{ALL} gene expression data. See
  \code{\link{topGOdata-class}} for code examples on how-to generate this object.
}

\examples{
data(GOdata)

## print the object
GOdata
}

\keyword{datasets}