\name{printGenes-methods}
\docType{methods}


\alias{printGenes-methods}
\alias{printGenes}
\alias{printGenes,topGOdata,character,character-method}
\alias{printGenes,topGOdata,character,missing-method}

\title{Summary for genes annotated to a GO term}

\description{
  Function to print summary for the top genes annotated to the specified
  GO term.
}


\section{Methods}{
\describe{

\item{}{ ~~describe this method here }
}}

\author{Adrian Alexa}

\seealso{
  \code{\link{classicScore-class}},
  \code{\link{groupStats-class}},
  \code{\link{getSigGroups-methods}}
}

\keyword{methods}

